+++ lib/domain/loan_particulars/models/loan_particulars.dart
@@ -11,6 +11,7 @@ class LoanParticulars with _$LoanParticulars {
     required VehicleDetails vehicleDetails,
     required LoanDetails loanDetails,
     required EMIDetails emiDetails,
+    double? fundedLoanAmount,
     double? ddAmount,
     double? downPayment,
   }) = _LoanParticulars;