+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # DataDex Changelog 
 
+## 30 April 2024
+
+### Features
+
+- Setup calculations of DD and downpayment based in loanscheme
+- Setup option to set loan scheme and set funded charges
+- Setup option to add landmark in address
+
+
 ## 28 March 2024
 
 ### Bug Fixes